+++ Release_Notes.md
@@ -4,6 +4,7 @@
 ### New Features
 
 ### Improvements
+- Removed centralManagement from the BClientChoices constructor ... it's not needed
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure